create sandbox for student activity
move sample html and css to ../sandbox from module
move (already ignored) Website and Alice to ..sandbox
make top-level .gitignore.

Note that a directory called sandbox will be ignored at any level, so
there could be one in every module.  That would probably be a good
idea in the future.