Add credits in jpeg comments to image files
This change adds "File source: ..." and "BY: ..." comments, which should
be enough to (minimally) satisfy the CC attribution requirements for the
student-created website, given that it's not likely to be on the web at
all.  Attribution should also be given in either one of the exercises,
the module summary, or both.